requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
branca==0.8.2
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.0
cycler==0.12.1
folium==0.20.0
fonttools==4.56.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.3
matplotlib==3.9.4
networkx==3.2.1
numpy==2.0.2
packaging==24.2
pillow==11.1.0
pygame==2.6.1
pyparsing==3.2.1
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-ev3dev2==2.1.0.post1
requests==2.32.5
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
urllib3==2.5.0
vboxapi==1.0
xyzservices==2025.4.0
zipp==3.21.0
